requirements.txt
python-dotenv
anthropic
pydantic
tqdm